Split up CLI
